Fixed url function, still issues with imports. Test cases added.
