+++ pdfminer/runlength.py
@@ -7,22 +7,6 @@
 #
 
 def rldecode(data):
-    r"""
-    RunLength decoder (Adobe version) implementation based on PDF Reference
-    version 1.4 section 3.3.4:
-        The RunLengthDecode filter decodes data that has been encoded in a
-        simple byte-oriented format based on run length. The encoded data
-        is a sequence of runs, where each run consists of a length byte
-        followed by 1 to 128 bytes of data. If the length byte is in the
-        range 0 to 127, the following length + 1 (1 to 128) bytes are
-        copied literally during decompression. If length is in the range
-        129 to 255, the following single byte is to be copied 257 - length
-        (2 to 128) times during decompression. A length value of 128
-        denotes EOD.
-    >>> s = b'\x05123456\xfa7\x04abcde\x80junk'
-    >>> rldecode(s)
-    b'1234567777777abcde'
-    """
     decoded = b''
     i = 0
     while i < len(data):
@@ -41,9 +25,3 @@ def rldecode(data):
             decoded += run
             i = (i + 1) + 1
     return decoded
-
-
-if __name__ == '__main__':
-    import doctest
-
-    print('pdfminer.runlength', doctest.testmod())
